requirements.txt
Django~=5.1.6
pillow
djangorestframework